requirements.txt
numpy==1.24.4
pandas==2.0.3
pdfminer.six==20221105
python-docx==0.8.11
Pillow==10.0.0
pytesseract==0.3.10
scikit-learn==1.3.0
